Followed advanced programmers advices and cleaned up the code

+++ dummy_sorter.hpp
@@ -8,7 +8,7 @@ namespace sorter {
     class DummySorter
     {
     private:
-        int _memoryLimit;
+        size_t _memoryLimit;
         std::string _inputName, _outputName, _tmdDir;
 
     public:
